+++ fastapi_admin/models.py
@@ -15,7 +15,7 @@ class AbstractUser(Model):
         abstract = True
 
 
-class Permission(Model):
+class AbstractPermission(Model):
     label = fields.CharField(max_length=50)
     model = fields.CharField(max_length=50)
     action: enums.PermissionAction = fields.IntEnumField(
@@ -25,12 +25,18 @@ class Permission(Model):
     def __str__(self):
         return self.label
 
+    class Meta:
+        abstract = True
 
-class Role(Model):
+
+class AbstractRole(Model):
     label = fields.CharField(max_length=50)
     users = fields.ManyToManyField("models.User")
 
-    permissions: fields.ManyToManyRelation[Permission] = fields.ManyToManyField("models.Permission")
+    permissions = fields.ManyToManyField("models.Permission")
 
     def __str__(self):
         return self.label
+
+    class Meta:
+        abstract = True